refactor: Move `getElementFromSelector` & `getSelectorFromElement`
Move `getElementFromSelector` & getSelectorFromElement` inside selector-engine.js, in order to use SelectorEngine methods, avoiding raw querySelector usage

Feature: add `getMultipleElementsFromSelector` helper
